refactor: extract gateway admin issue reports
