+++ crates/dcc-mcp-gateway-admin/Cargo.toml
@@ -14,6 +14,7 @@ workspace = true
 
 [dependencies]
 axum.workspace = true
+chrono = { version = "0.4", features = ["serde"] }
 parking_lot.workspace = true
 serde.workspace = true
 serde_json.workspace = true
